% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blazr-package.R
\docType{package}
\name{blazr-package}
\alias{blazr}
\alias{blazr-package}
\title{blazr: Blazingly Fast Multi-Threading Interface for R Using Rust}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A tool for performing multi-threaded computations in R, powered by Rust.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-staceans/blazr}
  \item \url{https://r-staceans.github.io/blazr/}
  \item Report bugs at \url{https://github.com/r-staceans/blazr/issues}
}

}
\author{
\strong{Maintainer}: Jackson Hoffart \email{jackson.hoffart@gmail.com} (\href{https://orcid.org/0000-0002-8600-5042}{ORCID})

Authors:
\itemize{
  \item Bolívar Aponte Rolón (\href{https://orcid.org/0000-0002-2544-4551}{ORCID})
}

}
\keyword{internal}
