% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-wrappers.R
\name{sum_with_threads}
\alias{sum_with_threads}
\title{Calculate the sum of a vector of integers using multiple threads.}
\usage{
sum_with_threads(x, n)
}
\arguments{
\item{x}{A vector of integers to sum over.}

\item{n}{The number of threads used to compute this calculation (int).}
}
\value{
The sum of all elements of the input vector.
}
\description{
Calculate the sum of a vector of integers using multiple threads.
}
